backend/requirements.txt
fastapi[all]
uvicorn[standard]
google-generativeai
pydantic-settings
python-multipart
python-dotenv
pathlib
langchain
langchain-google-genai
langgraph
pypdf
docx2txt